UI/Controller 분리
